Update ContentModel.java ---> Scroll Event
Scrolling allows a zoom in and zoom out. In the coordinates of clas12 this was extremely slow and awkward as the forward carriage is in the +z hemisphere. I propose a change in the scroll event to give faster zoom and the ability for the camera to go into the +z facing out, without complicated rotations first. I would also argue the ZoomEvent and Translate camera (press on mouse wheel then mouse movement) should be changed to faster as well.